+++ platform/src/codatplatform/models/operations/get_integration.py
@@ -10,7 +10,7 @@ from typing import Optional
 @dataclasses.dataclass
 class GetIntegrationRequest:
     platform_key: str = dataclasses.field(metadata={'path_param': { 'field_name': 'platformKey', 'style': 'simple', 'explode': False }})
-    r"""A unique 4-letter key to represent a platform in each integration. View [accounting](https://docs.codat.io/integrations/accounting/overview#platform-keys), [banking](https://docs.codat.io/integrations/banking/overview#platform-keys), and [commerce](https://docs.codat.io/integrations/commerce/overview#platform-keys) platform keys."""
+    r"""A unique 4-letter key to represent a platform in each integration."""