+++ platform/src/codatplatform/sdkconfiguration.py
@@ -2,6 +2,7 @@
 
 
 import requests as requests_http
+from ._hooks import SDKHooks
 from .utils import utils
 from .utils.retries import RetryConfig
 from codatplatform.models import shared
@@ -23,10 +24,11 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '3.3.0'
-    gen_version: str = '2.257.2'
-    user_agent: str = 'speakeasy-sdk/python 3.3.0 2.257.2 3.0.0 codat-platform'
+    sdk_version: str = '3.4.0'
+    gen_version: str = '2.277.0'
+    user_agent: str = 'speakeasy-sdk/python 3.4.0 2.277.0 3.0.0 codat-platform'
     retry_config: RetryConfig = None
+    _hooks: SDKHooks = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]:
         if self.server_url:
@@ -35,3 +37,7 @@ class SDKConfiguration:
             self.server_idx = 0
 
         return SERVERS[self.server_idx], {}
+
+
+    def get_hooks(self) -> SDKHooks:
+        return self._hooks